backend/requirements.txt
# Backend framework
fastapi>=0.115.2,<0.116
uvicorn[standard]>=0.30.6,<0.31

# Database
sqlmodel>=0.0.22,<0.1
SQLAlchemy>=2.0.35,<2.1

# Data validation
pydantic>=2.9.2,<3.0
pydantic-settings>=2.6.0,<3.0

# Environment & security
python-dotenv>=1.0.1,<2.0
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0

# HTTP clients
requests>=2.32.3,<3.0
beautifulsoup4>=4.12.3,<5.0
httpx>=0.27.2,<1.0

# AI / reasoning
langchain>=0.3.2,<0.4
langgraph>=0.6.7,<0.7
langchain-google-vertexai>=2.0

# Web scraping
crawl4ai[pdf]>=0.3.0,<1.0
playwright>=1.40.0,<2.0
